+++ Apps/Contoso.Forms.Puppet/Contoso.Forms.Puppet/Constants.cs
@@ -11,6 +11,6 @@ namespace Contoso.Forms.Puppet
         public const string UpdateTrackKey = "update-track";
         public const string AutomaticUpdateCheckKey = "automatic-update-check";
         public const string StorageMaxSize = "storage-max-size";
-        public const string EnableManualSessionTracker = "enable-manual-session-generation";
+        public const string EnableManualSessionTracker = "enable-manual-session-tracker";
     }
 }